updated the type of a function from float to int, since it was being returned to an int value

+++ include/calculator.h
@@ -19,7 +19,7 @@ private:
 	char last_op;
 
 	void handle_button_event(union blue_trellis::button_event press);
-	float calc_total(float num1, float num2, char op);
+	int calc_total(float num1, float num2, char op);
 	void update_lcd(std::string num);
 	std::string num_to_string(float num);
 	std::string num_to_string(int num);
